refactor(auth_handlers): split 536 LOC en session/oauth/totp
- auth_handlers/session.rs (357 LOC): login/register/logout/refresh + password reset + patch_locale + make_session helpers
- auth_handlers/oauth.rs (90 LOC): auth_oauth_start + auth_oauth_callback
- auth_handlers/totp.rs (86 LOC): HOTP/TOTP helpers + api_2fa_verify
- mod.rs (11 LOC): re-exports globaux, préserve super::auth_handlers::*

Zéro changement de comportement. main.rs continue d'utiliser pub use auth_handlers::*.
